- Rebase - source out visualizations - source out perfdata parsing/handling - collect perfdata also for scorep processes - abstract monitor for parallel setups
