+++ src/pyperf_jupyter/kernel.py
@@ -1,38 +1,26 @@
-from ipykernel.ipkernel import IPythonKernel
-import sys
+import json
 import os
+import re
+import subprocess
+import sys
+import importlib
 
-# Create interactive widgets
-from ipywidgets import interact, interactive, fixed, interact_manual
-import ipywidgets as widgets
-from IPython.display import display
-from pyperf_jupyter.userpersistence import PersHelper, scorep_script_name
 from enum import Enum
 from textwrap import dedent
-
-import pandas as pd
-from functools import partial
-
 from statistics import mean
-import json
-
-import subprocess
-import re
-import json
-import time
-import pickle
-import codecs
-import matplotlib.pyplot as plt
-import matplotlib.colors as mcolors
-from itables import show
 import pandas as pd
-from datetime import datetime
+from ipykernel.ipkernel import IPythonKernel
+from itables import show
 from pyperf_jupyter.userpersistence import PersHelper, scorep_script_name
+from pyperf_jupyter.userpersistence import magics_cleanup
+
+from pyperf_jupyter.perfdatahandler import PerformanceDataHandler
+import pyperf_jupyter.visualization as perfvis
+import pyperf_jupyter.parallel_monitor.slurm_monitor as slurm_monitor
 
 PYTHON_EXECUTABLE = sys.executable
 READ_CHUNK_SIZE = 8
 userpersistence_token = "pyperf_jupyter.userpersistence"
-scorep_script_name = "scorep_script.py"
 jupyter_dump = "jupyter_dump.pkl"
 subprocess_dump = "subprocess_dump.pkl"
 
@@ -70,9 +58,6 @@ class PyPerfKernel(IPythonKernel):
 
         self.blacklist_prefixes = ['%lsmagic']
 
-        self.code_history = []
-        self.performance_data_history = []
-
         self.scorep_binding_args = []
         self.scorep_env = {}
 
@@ -91,11 +76,10 @@ class PyPerfKernel(IPythonKernel):
         self.writefile_scorep_binding_args = []
         self.writefile_multicell = False
 
-        self.slurm_nodelist = self.parse_slurm_nodelist(os.environ.get("SLURM_NODELIST", ''))
-        if len(self.slurm_nodelist) <= 1:
-            self.slurm_nodelist = None
         # will be set to True as soon as GPU data is received
         self.gpu_avail = False
+        self.perfdata_handler = PerformanceDataHandler()
+        self.nodelist = self.perfdata_handler.get_nodelist()
 
     def cell_output(self, string, stream='stdout'):
         """
@@ -139,6 +123,28 @@ class PyPerfKernel(IPythonKernel):
             self.cell_output(f'KernelWarning: Currently in {self.mode}, command ignored.', 'stderr')
         return self.standard_reply()
 
+    def set_perfmonitor(self, code):
+        """
+        Read the perfmonitor and try to select it.
+        """
+        if self.mode == KernelMode.DEFAULT:
+            monitor = code.split('\n')[1]
+            if monitor in {'local', "localhost", "LOCAL", "LOCALHOST"}:
+                self.cell_output('Selected local monitor. No parallel monitoring.')
+            else:
+                try:
+                    self.perfdata_handler.set_monitor(monitor)
+                    self.nodelist = self.perfdata_handler.get_nodelist()
+                    if len(self.nodelist) <= 1:
+                        self.nodelist = None
+                        self.cell_output('Found monitor: ' + str(monitor) + ' but no nodelist, using local setup. ')
+                    else:
+                        self.cell_output('Selected monitor: ' + str(monitor) + ' and got nodes: ' + str(self.nodelist))
+                except:
+                    self.cell_output(f'Error setting monitor', 'stderr')
+        else:
+            self.cell_output(f'KernelWarning: Currently in {self.mode}, command ignored.', 'stderr')
+        return self.standard_reply()
 
     def set_scorep_env(self, code):
         """
@@ -294,305 +300,12 @@ class PyPerfKernel(IPythonKernel):
         self.pershelper.postprocess()
         self.cell_output(error_message, 'stderr')
 
-
-    def parse_slurm_nodelist(self, nodelist_string):
-        nodes = []
-        # Regular expression to match patterns like 'node[1-3],node5,node7'
-        pattern = r'(\w+\[\d+-\d+\]|\w+|\d+)'
-        matches = re.findall(pattern, nodelist_string)
-
-        for match in matches:
-            # Check if the match contains a range
-            if '-' in match:
-                node_range = match.split('[')
-                prefix = node_range[0]
-                start, end = map(int, re.findall(r'\d+', node_range[1]))
-                nodes.extend([f"{prefix}{i}" for i in range(start, end + 1)])
-            else:
-                nodes.append(match)
-
-        return nodes
-
-    def get_perfdata_aggregated(self):
-        perfdata_aggregated = []
-        # for each node, initialize the list
-        # index 0 in next row indicates, we are just checking the very first cell that was executed
-        '''
-        each line in performance_data_history represents a cell (code) that was executed
-        each cell then looks like that:
-        [raw data node0,
-             raw data node1,
-             raw data noden,
-             CPU Mean across nodes,
-             Mem Mean across nodes,
-             IO OPS READ Mean across nodes,
-             IO OPS WRITE Mean across nodes,
-             IO Bytes READ Mean across nodes,
-             IO Bytes WRITE Mean across nodes,
-             GPU Util Mean across nodes,
-             GPU Mem Mean across nodes]
-        '''
-        for node in range(0, len(self.performance_data_history[0])-8):
-            perfdata_init = self.performance_data_history[0][node]
-            # for each cpu and gpu, we need an empty list to be filled
-            cpu_util = [[] for _ in perfdata_init[0]] if perfdata_init[0] else [[]]
-            gpu_util = [[] for _ in perfdata_init[2]] if perfdata_init[2] else [[]]
-            gpu_mem = [[] for _ in perfdata_init[2]] if perfdata_init[3] else [[]]
-            mem_util = []
-            io_ops_read = []
-            io_ops_write = []
-            io_bytes_read = []
-            io_bytes_write = []
-            perfdata_aggregated.append([cpu_util, mem_util, io_ops_read, io_ops_write, io_bytes_read, io_bytes_write,
-                                        gpu_util, gpu_mem])
-
-        # for each cell/code (called perfdata)...
-        for perfdata in self.performance_data_history:
-            # for each node in this cell/code
-            for node in range(0,len(perfdata)-8):
-                for cpu_index in range(0, len(perfdata[node][0])):
-                    perfdata_aggregated[node][0][cpu_index].extend(perfdata[node][0][cpu_index])
-                perfdata_aggregated[node][1].extend(perfdata[node][1])
-                perfdata_aggregated[node][2].extend(perfdata[node][2])
-                perfdata_aggregated[node][3].extend(perfdata[node][3])
-                perfdata_aggregated[node][4].extend(perfdata[node][4])
-                perfdata_aggregated[node][5].extend(perfdata[node][5])
-                for gpu_index in range(0, len(perfdata[node][6])):
-                    perfdata_aggregated[node][6][gpu_index].extend(perfdata[node][6][gpu_index])
-                for gpu_index in range(0, len(perfdata[node][7])):
-                    perfdata_aggregated[node][7][gpu_index].extend(perfdata[node][7][gpu_index])
-
-        return perfdata_aggregated
-
-
-    def plot_graph(self, ax, metric, perfdata):
-        # first 0 means first node
-        ax.clear()  # Clear previous plot
-        #generate scale
-        x_scale = [x for x in range(0, 2 * len(perfdata[0][0][-3]), int(os.environ.get("PYPERF_REPORT_FREQUENCY", 2)))]
-        if metric == 'CPU Util (Min/Max/Mean)':
-            ax.plot(x_scale, perfdata[0][0][-3], label='Mean', color=(0.20,0.47,1.00))
-            ax.plot(x_scale, perfdata[0][0][-2], label='Max', color=(0.20,0.47,1.00, 0.3))
-            ax.plot(x_scale, perfdata[0][0][-1], label='Min', color=(0.20,0.47,1.00, 0.3))
-            ax.set_ylabel('Util (%)')
-        elif metric == 'CPU Cores (Raw)':
-            for cpu_index in range(0, len(perfdata[0][0])-3):
-                ax.plot(x_scale, perfdata[0][0][cpu_index], label="CPU" + str(cpu_index))
-            ax.set_ylabel('Util (%)')
-        elif metric == 'Mem':
-            ax.plot(x_scale, perfdata[0][1], label='Value', color=(0.12,0.70,0.00))
-            ax.set_ylabel('Util (%)')
-        elif metric == 'IO Ops':
-            ax.plot(x_scale, perfdata[0][2], label='IO Read', color=(1.00,1.00,0.10))
-            ax.plot(x_scale, perfdata[0][3], label='IO Write', color=(1.00,0.50,0.00))
-            ax.set_ylabel('Ops')
-        elif metric == 'IO Bytes':
-            ax.plot(x_scale, perfdata[0][4], label='IO Read', color=(0.50,0.50,0.00))
-            ax.plot(x_scale, perfdata[0][5], label='IO Write', color=(0.50,0.25,0.00))
-            ax.set_ylabel('Bytes')
-        elif metric == 'GPU Util':
-            ax.plot(x_scale, perfdata[0][6][-3], label='Mean', color=(0.90,0.30,0.00))
-            ax.plot(x_scale, perfdata[0][6][-2], label='Max', color=(0.90,0.30,0.00, 0.3))
-            ax.plot(x_scale, perfdata[0][6][-1], label='Min', color=(0.90,0.30,0.00, 0.3))
-            ax.set_ylabel('Util (%)')
-        elif metric == 'GPU Mem':
-            ax.plot(x_scale, perfdata[0][7][-3], label='Mean', color=(1.00,0.40,1.00))
-            ax.plot(x_scale, perfdata[0][7][-2], label='Max', color=(1.00,0.40,1.00, 0.3))
-            ax.plot(x_scale, perfdata[0][7][-1], label='Min', color=(1.00,0.40,1.00, 0.3))
-            ax.set_ylabel('Util (%)')
-
-        ax.set_title(f'{metric}')
-        ax.set_xlabel('Time (s)')
-        ax.legend()
-        ax.grid(True)
-
-    def plot_with_dropdowns(self, metrics, perfdata, metric_start):
-        # Create subplots in a 1x2 grid
-        fig, axes = plt.subplots(1, 2, figsize=(10, 3))
-        dropdowns = []
-
-        # Plot data and create dropdowns for each subplot
-        for i, ax in enumerate(axes):
-            self.plot_graph(ax, metrics[i+metric_start], perfdata)
-
-            # Create dropdown widget for the current subplot
-            dropdown = widgets.Dropdown(options=metrics, description='Metric:', value=metrics[i+metric_start])
-            dropdown.observe(lambda change, ax=ax: self.plot_graph(ax, change['new'], perfdata), names='value')
-
-            # Add dropdown to list
-            dropdowns.append(dropdown)
-            # Display dropdowns and plots
-
-        display(widgets.HBox(dropdowns, layout=widgets.Layout(margin='0 0', padding='0px 15%')))
-        plt.tight_layout()
-        plt.show()
-
-
-    def draw_performance_graph(self, perfdata):
-
-        if self.slurm_nodelist:
-            nodelist = self.slurm_nodelist
-            nodelist.insert(0, 'All')
-            dropdown = widgets.Dropdown(
-                options=nodelist,
-                value='All',
-                description='Number:',
-                disabled=False,
-            )
-            display(dropdown)
-
-        # Dropdown widget
-        metrics = ['CPU Util (Min/Max/Mean)', 'CPU Cores (Raw)', 'Mem', 'IO Ops', 'IO Bytes']
-        if self.gpu_avail:
-            metrics.extend(["GPU Util", "GPU Mem"])
-
-        button = widgets.Button(description="Add Display")
-        output = widgets.Output()
-
-        display(button, output)
-
-        def on_button_clicked(b):
-            with output:
-                self.plot_with_dropdowns(metrics, perfdata, 0)
-
-        button.on_click(on_button_clicked)
-
-        self.plot_with_dropdowns(metrics, perfdata, 0)
-        if self.gpu_avail:
-            self.plot_with_dropdowns(metrics, perfdata, 2)
-
-
-
-
-    def compute_mean_across_nodes(self, key1, key2, metrics_across_nodes):
-        # use the first list
-        mean_across_nodes = []
-        for i in range(0,len(metrics_across_nodes[key1][key2][0])):
-            mean = metrics_across_nodes[key1][key2][0][i]
-            # check the values on the other lists, compute the mean and add
-            for j in range(1,len(metrics_across_nodes[key1][key2])):
-                mean+=metrics_across_nodes[key1][key2][j][i]
-            mean_across_nodes.append(mean/len(metrics_across_nodes[key1][key2]))
-        return mean_across_nodes
-
-    def report_perfdata(self, stdout_data_node, duration, nodelist):
-
-        # might be that parent process pushes to stdout and that output gets reversed
-        # couldn't find a better way than waiting for 1s and hoping that outputs are in correct order in the queue now
-        time.sleep(1)
-
-        # parse the performance data
-        performance_data_nodes = []
-        metrics_across_nodes = {'CPU': {'MEANS': [],
-                                'MAXS': [],
-                                'MINS': []},
-                        'MEM': {'MEANS': [],
-                                'MAXS': [],
-                                'MINS': []},
-                        'GPU_UTIL': {'MEANS': [],
-                                'MAXS': [],
-                                'MINS': []},
-                        'GPU_MEM': {'MEANS': [],
-                                     'MAXS': [],
-                                     'MINS': []},
-                        'IO_OPS_READ': {'MEANS': [],
-                                    'MAXS': [],
-                                    'MINS': []},
-                        'IO_OPS_WRITE': {'MEANS': [],
-                                        'MAXS': [],
-                                        'MINS': []},
-                        'IO_BYTES_READ': {'MEANS': [],
-                                         'MAXS': [],
-                                         'MINS': []},
-                        'IO_BYTES_WRITE': {'MEANS': [],
-                                          'MAXS': [],
-                                          'MINS': []},
-                         }
-        for stdout_data in stdout_data_node:
-            stdout_data = stdout_data.split(b"\n\n")
-            # init CPU and GPU lists
-            mem_util = []
-            io_ops_read = []
-            io_ops_write = []
-            io_bytes_read = []
-            io_bytes_write = []
-            if stdout_data[0]:
-                init_data = pickle.loads(codecs.decode(stdout_data[0], "base64"))
-                cpu_util = [[] for _ in init_data[0]]
-                gpu_util = [[] for _ in init_data[2]]
-                gpu_mem = [[] for _ in init_data[3]]
-
-                # add empty lists for mean, max, min
-                for i in range(0,3):
-                    cpu_util.append([])
-                    gpu_util.append([])
-                    gpu_mem.append([])
-            else:
-                return False
-
-            for line in stdout_data:
-                if line == b'':
-                    continue
-                perf_data = pickle.loads(codecs.decode(line, "base64"))
-
-                for cpu in range(0, len(perf_data[0])):
-                    cpu_util[cpu].append(perf_data[0][cpu])
-
-                last_measurements = [cpu_list[-1] for cpu_list in cpu_util[:-3]]
-                cpu_util[-3].append(mean(last_measurements))
-                cpu_util[-2].append(max(last_measurements))
-                cpu_util[-1].append(min(last_measurements))
-
-
-                mem_util.append(perf_data[1])
-                #0 read counts, 1 write counts, 2 read bytes, 3 write bytes
-                io_ops_read.append(perf_data[4][0])
-                io_ops_write.append(perf_data[4][1])
-                io_bytes_read.append(perf_data[4][2])
-                io_bytes_write.append(perf_data[4][3])
-                for gpu in range(0, len(perf_data[2])):
-                    gpu_util[gpu].append(perf_data[2][gpu])
-                    gpu_mem[gpu].append(perf_data[3][gpu])
-
-                last_measurements = [gpu_list[-1] for gpu_list in gpu_util[:-3]]
-                if last_measurements:
-                    gpu_util[-3].append(mean(last_measurements))
-                    gpu_util[-2].append(max(last_measurements))
-                    gpu_util[-1].append(min(last_measurements))
-
-                last_measurements = [gpu_list[-1] for gpu_list in gpu_mem[:-3]]
-                if last_measurements:
-                    gpu_mem[-3].append(mean(last_measurements))
-                    gpu_mem[-2].append(max(last_measurements))
-                    gpu_mem[-1].append(min(last_measurements))
-
-            # TODO: discuss whether Means of Maxs are relevant or should we consider Mean of Means only?
-            metrics_across_nodes["CPU"]["MEANS"].append(cpu_util[-3])
-            metrics_across_nodes["MEM"]["MEANS"].append(mem_util)
-            metrics_across_nodes["IO_OPS_READ"]["MEANS"].append(io_ops_read)
-            metrics_across_nodes["IO_OPS_WRITE"]["MEANS"].append(io_ops_write)
-            metrics_across_nodes["IO_BYTES_READ"]["MEANS"].append(io_bytes_read)
-            metrics_across_nodes["IO_BYTES_WRITE"]["MEANS"].append(io_bytes_write)
-            metrics_across_nodes["GPU_UTIL"]["MEANS"].append(gpu_util[-3])
-            metrics_across_nodes["GPU_MEM"]["MEANS"].append(gpu_mem[-3])
-            performance_data_nodes.append([cpu_util, mem_util, io_ops_read, io_ops_write, io_bytes_read, io_bytes_write, gpu_util, gpu_mem])
-
-        # compute the mean, max, min across the nodes using pure mem_util lists, the mean, max and min lists of the other sensors
-        performance_data_nodes.append(self.compute_mean_across_nodes("CPU", "MEANS", metrics_across_nodes))
-        performance_data_nodes.append(self.compute_mean_across_nodes("MEM", "MEANS", metrics_across_nodes))
-        performance_data_nodes.append(self.compute_mean_across_nodes("IO_OPS_READ", "MEANS", metrics_across_nodes))
-        performance_data_nodes.append(self.compute_mean_across_nodes("IO_OPS_WRITE", "MEANS", metrics_across_nodes))
-        performance_data_nodes.append(self.compute_mean_across_nodes("IO_BYTES_READ", "MEANS", metrics_across_nodes))
-        performance_data_nodes.append(self.compute_mean_across_nodes("IO_BYTES_WRITE", "MEANS", metrics_across_nodes))
-        performance_data_nodes.append(self.compute_mean_across_nodes("GPU_UTIL", "MEANS", metrics_across_nodes))
-        performance_data_nodes.append(self.compute_mean_across_nodes("GPU_MEM", "MEANS", metrics_across_nodes))
-
-        self.performance_data_history.append(performance_data_nodes)
+    def report_perfdata(self, performance_data_nodes, duration):
 
         # print the performance data
         report_trs = int(os.environ.get("PYPERF_REPORTS_MIN", 2))
 
-        #just count the number of memory measurements to decide whether we want to print the information
+        # just count the number of memory measurements to decide whether we want to print the information
         if len(performance_data_nodes[0][1]) > report_trs:
 
             self.cell_output("\n----Performance Data----\n", 'stdout')
@@ -617,8 +330,8 @@ class PyPerfKernel(IPythonKernel):
 
             for idx, performance_data in enumerate(performance_data_nodes[:-8]):
 
-                if nodelist:
-                    self.cell_output("--NODE " + str(nodelist[idx]) + "--\n", 'stdout')
+                if self.nodelist:
+                    self.cell_output("--NODE " + str(self.nodelist[idx]) + "--\n", 'stdout')
 
                 cpu_util = performance_data[0]
                 mem_util = performance_data[1]
@@ -629,9 +342,8 @@ class PyPerfKernel(IPythonKernel):
                 gpu_util = performance_data[6]
                 gpu_mem = performance_data[7]
 
-
                 if cpu_util:
-                    #self.cell_output("--CPU Util--\n", 'stdout')
+                    # self.cell_output("--CPU Util--\n", 'stdout')
                     self.cell_output(
                         "\nCPU Util    \tAVG: " + "{:.2f}".format(
                             mean(cpu_util[-3])) + "\t MIN: " + "{:.2f}".format(
@@ -723,10 +435,6 @@ class PyPerfKernel(IPythonKernel):
                                      'stdout')
             '''
 
-            return True
-        else:
-            return False
-
     async def scorep_execute(self, code, silent, store_history=True, user_expressions=None,
                              allow_stdin=False, *, cell_id=None):
         """
@@ -760,6 +468,8 @@ class PyPerfKernel(IPythonKernel):
             self.scorep_binding_args + [scorep_script_name]
         proc_env = self.scorep_env.copy()
         proc_env.update({'PATH': os.environ['PATH'], 'PYTHONUNBUFFERED': 'x'}) # scorep path, subprocess observation
+
+        self.perfdata_handler.start_perfmonitor()
         proc = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, env=proc_env)
         
         # For memory mode jupyter_dump and jupyter_update must be awaited
@@ -794,6 +504,8 @@ class PyPerfKernel(IPythonKernel):
                     incomplete_line = ""
                 for line in lines:
                     self.cell_output(line)
+
+        performance_data_nodes, duration = self.perfdata_handler.end_perfmonitor(code)
 
         # In disk mode, subprocess already terminated after dumping persistence to file
         if self.pershelper.mode == 'disk':
@@ -837,6 +549,8 @@ class PyPerfKernel(IPythonKernel):
                 self.cell_output("KernelWarning: Instrumentation results were not saved locally.", 'stderr')
 
         self.pershelper.postprocess()
+        if performance_data_nodes:
+            self.report_perfdata(performance_data_nodes, duration)
         return self.standard_reply()
 
     async def do_execute(self, code, silent, store_history=False, user_expressions=None,
@@ -870,40 +584,40 @@ class PyPerfKernel(IPythonKernel):
             return self.standard_reply()
         '''
         if code.startswith('%%display_graph_for_last'):
-            self.draw_performance_graph(self.performance_data_history[-1])
+            perfvis.draw_performance_graph(self.nodelist, self.perfdata_handler.get_perfdata_history()[-1], self.gpu_avail)
             return self.standard_reply()
         elif code.startswith('%%display_graph_for_index'):
             if len(code.split(' '))==1:
                 self.cell_output("No index specified. Use: %%display_graph_for_index index", 'stdout')
             index = int(code.split(' ')[1])
-            if index>=len(self.performance_data_history):
-                self.cell_output("Tracked only "+ str(len(self.performance_data_history)) +" cells. This index is not available.")
+            if index>=len(self.perfdata_handler.get_perfdata_history()):
+                self.cell_output("Tracked only " + str(len(self.perfdata_handler.get_perfdata_history())) + " cells. This index is not available.")
             else:
-                self.draw_performance_graph(self.performance_data_history[index])
+                perfvis.draw_performance_graph(self.nodelist, self.perfdata_handler.get_perfdata_history()[index], self.gpu_avail)
             return self.standard_reply()
         elif code.startswith('%%display_graph_for_all'):
-            self.draw_performance_graph(self.get_perfdata_aggregated())
+            perfvis.draw_performance_graph(self.nodelist, self.perfdata_handler.get_perfdata_aggregated(), self.gpu_avail)
             return self.standard_reply()
 
         elif code.startswith('%%display_code_for_index'):
             if len(code.split(' '))==1:
                 self.cell_output("No index specified. Use: %%display_code_for_index index", 'stdout')
             index = int(code.split(' ')[1])
-            if index >= len(self.performance_data_history):
-                self.cell_output("Tracked only "+ str(len(self.performance_data_history)) +" cells. This index is not available.")
+            if index >= len(self.perfdata_handler.get_perfdata_history()):
+                self.cell_output("Tracked only " + str(len(self.perfdata_handler.get_perfdata_history())) + " cells. This index is not available.")
             else:
-                self.cell_output("Cell timestamp: " + str(self.code_history[index][0]) + "\n--\n", 'stdout')
-                self.cell_output(self.code_history[index][1], 'stdout')
+                self.cell_output("Cell timestamp: " + str(self.perfdata_handler.get_code_history()[index][0]) + "\n--\n", 'stdout')
+                self.cell_output(self.perfdata_handler.get_code_history()[index][1], 'stdout')
             return self.standard_reply()
         elif code.startswith('%%display_code_history'):
-            show(pd.DataFrame(self.code_history, columns=["timestamp", "code"]).reset_index())
+            show(pd.DataFrame(self.perfdata_handler.get_code_history(), columns=["timestamp", "code"]).reset_index())
             return self.standard_reply()
         elif code.startswith('%%perfdata_to_variable'):
             if len(code.split(' '))==1:
                 self.cell_output("No variable to export specified. Use: %%perfdata_to_variable myvar", 'stdout')
             else:
                 varname = code.split(' ')[1]
-                await super().do_execute(f"{varname}={self.performance_data_history}", silent=True)
+                await super().do_execute(f"{varname}={self.perfdata_handler.get_perfdata_history()}", silent=True)
                 self.cell_output("Exported performance data to " + str(varname) + " variable", 'stdout')
             return self.standard_reply()
         elif code.startswith('%%perfdata_to_json'):
@@ -912,12 +626,14 @@ class PyPerfKernel(IPythonKernel):
             else:
                 filename = code.split(' ')[1]
                 with open(f'{filename}_perfdata.json', 'w') as f:
-                    json.dump(self.performance_data_history, default=str, fp=f)
+                    json.dump(self.perfdata_handler.get_perfdata_history(), default=str, fp=f)
                 with open(f'{filename}_code.json', 'w') as f:
-                    json.dump(self.code_history, default=str, fp=f)
+                    json.dump(self.perfdata_handler.get_code_history(), default=str, fp=f)
                 self.cell_output("Exported performance data to " +
                                  str(filename) + "_perfdata.json and " + str(filename) + "_code.json", 'stdout')
             return self.standard_reply()
+        elif code.startswith('%%set_perfmonitor'):
+            return self.set_perfmonitor(code)
         elif code.startswith('%%scorep_env'):
             return self.set_scorep_env(code)
         elif code.startswith('%%scorep_python_binding_arguments'):
@@ -928,7 +644,6 @@ class PyPerfKernel(IPythonKernel):
             return self.enable_multicellmode()
         elif code.startswith('%%abort_multicellmode'):
             return self.abort_multicellmode()
-
         elif code.startswith('%%finalize_multicellmode'):
             # Cannot be put into a separate function due to tight coupling between do_execute and scorep_execute
             if self.mode == KernelMode.MULTICELL:
@@ -961,53 +676,16 @@ class PyPerfKernel(IPythonKernel):
         else:
             if self.mode == KernelMode.DEFAULT:
                 self.pershelper.parse(magics_cleanup(code), 'jupyter')
-                return await super().do_execute(code, silent, store_history, user_expressions, allow_stdin, cell_id=cell_id)
+                self.perfdata_handler.start_perfmonitor()
+                parent_ret = await super().do_execute(code, silent, store_history, user_expressions, allow_stdin, cell_id=cell_id)
+                performance_data_nodes, duration = self.perfdata_handler.end_perfmonitor(code)
+                if performance_data_nodes:
+                    self.report_perfdata(performance_data_nodes, duration)
+                return parent_ret
             elif self.mode == KernelMode.MULTICELL:
                 return self.append_multicellmode(magics_cleanup(code))
             elif self.mode == KernelMode.WRITEFILE:
                 return self.append_writefile(magics_cleanup(code), explicit_scorep=False)
-            self.pershelper.parse(code, 'jupyter')
-            # for code without scorep, we are interested in performance data
-            # start a subprocess that tracks the performance data
-            proc_env = os.environ.copy()
-            proc_env.update({'PYTHONUNBUFFERED': 'x'})
-            if self.slurm_nodelist:
-                # SLURM is available, use srun to get output for all the nodes
-                # until first \n we would like to get the hostname
-                perf_proc = subprocess.Popen(["srun", "hostname", "&&", "echo", "PERFDATA", "&&",
-                                              sys.executable, "-m" "perfmonitor"],
-                                                     stdout=subprocess.PIPE, stderr=subprocess.STDOUT, env=proc_env)
-            else:
-                # SLURM is not available, sorry we support only one node
-                perf_proc = subprocess.Popen([sys.executable, "-m" "perfmonitor"], stdout=subprocess.PIPE,
-                                             stderr=subprocess.STDOUT, env=proc_env)
-
-
-            start = time.perf_counter()
-            parent_ret = await super().do_execute(code, silent, store_history, user_expressions, allow_stdin,
-                                                  cell_id=cell_id)
-            duration = time.perf_counter() - start
-
-            perf_proc.kill()
-            output, _ = perf_proc.communicate()
-
-            stdout_data_node = []
-            '''
-            if self.slurm_nodelist:
-                output.split(b"PERFDATA")
-                for proc in node_subproc:
-                    proc.kill()
-                    output, _ = proc.communicate()
-                    stdout_data_node.append(output)
-            else:
-                stdout_data_node.append(output)
-            '''
-            stdout_data_node.append(output)
-
-            # parse the performance data from the stdout pipe of the subprocess and print the performance data
-            if self.report_perfdata(stdout_data_node, duration, self.slurm_nodelist):
-                self.code_history.append([datetime.now(), code])
-            return parent_ret
 
     def do_shutdown(self, restart):
         self.pershelper.postprocess()
