+++ src/pyperf_jupyter/parallel_monitor/abstract_monitor.py
@@ -1,0 +1,16 @@
+from abc import ABC, abstractmethod
+
+
+class AbstractMonitor(ABC):
+
+    @abstractmethod
+    def parse_nodelist(self):
+        pass
+
+    @abstractmethod
+    def start_monitor(self):
+        pass
+
+    @abstractmethod
+    def end_monitor(self):
+        pass
